requirements.txt
fastapi==0.88.0
pydantic==1.10.4
uvicorn==0.20.0
mangum==0.17.0
#psycopg2==2.9.5
# If psycopg2 does not work for your local environment, just uncomment line below and comment line above
psycopg2-binary
sqlalchemy==1.4.41
sqlmodel==0.0.8
fastapi-auth0==0.3.2
fastapi-pagination==0.12.3
python-dotenv==0.21.1
pyhumps==3.8.0
requests==2.28.2
python-multipart==0.0.5

# Used only for dev (test, linting)
black==22.12.0
httpx==0.23.3
pytest==7.2.1
coverage==7.1.0
